clean up cryptography-core
