+++ cryptography-core/src/asset_proofs/macros.rs
@@ -74,3 +74,21 @@ macro_rules! define_sub_assign_variants {
         }
     };
 }
+
+/// That `ensure` does not transform into a string representation like `failure::ensure` is doing.
+#[allow(unused_macros)]
+macro_rules! ensure {
+    ($predicate:expr, $context_selector:expr) => {
+        if !$predicate {
+            return Err($context_selector.into());
+        }
+    };
+}
+
+/// Helper macro to assert that `predicate` is an `Error::from( $err)`.
+#[allow(unused_macros)]
+macro_rules! assert_err {
+    ($predicate:expr, $err:expr) => {
+        assert_eq!($predicate.expect_err("Error expected").kind(), &$err);
+    };
+}
